requirements.txt
RPi.GPIO
gpiod
flask
numpy
pandas
spidev